feat: update api protoc version to v2.0.118 (#3326)

+++ scheduler/scheduling/scheduling.go
@@ -628,7 +628,7 @@ func ConstructSuccessNormalTaskResponse(candidateParents []*resource.Peer) *sche
 			Application:         &candidateParent.Task.Application,
 			FilteredQueryParams: candidateParent.Task.FilteredQueryParams,
 			RequestHeader:       candidateParent.Task.Header,
-			PieceLength:         uint32(candidateParent.Task.PieceLength),
+			PieceLength:         uint64(candidateParent.Task.PieceLength),
 			ContentLength:       uint64(candidateParent.Task.ContentLength.Load()),
 			PieceCount:          uint32(candidateParent.Task.TotalPieceCount.Load()),
 			SizeScope:           candidateParent.Task.SizeScope(),
